fix: channel API

+++ supabase/functions/_backend/public/channel.ts
@@ -25,8 +25,10 @@ interface GetDevice {
 }
 
 export async function get(c: Context, body: GetDevice, apikey: Database['public']['Tables']['apikeys']['Row']): Promise<Response> {
-  if (!body.app_id || !(await hasAppRight(c, body.app_id, apikey.user_id, 'read')))
+  if (!body.app_id || !(await hasAppRight(c, body.app_id, apikey.user_id, 'read'))) {
+    console.log('You can\'t access this app', body.app_id)
     return c.json({ status: 'You can\'t access this app', app_id: body.app_id }, 400)
+  }
 
   // get one channel or all channels
   if (body.channel) {
@@ -43,8 +45,8 @@ export async function get(c: Context, body: GetDevice, apikey: Database['public'
         disableAutoUpdateUnderNative,
         disableAutoUpdate,
         allow_device_self_set,
-        is_emulator,
-        is_prod,
+        allow_emulator,
+        public,
         version (
           name,
           id
@@ -53,8 +55,10 @@ export async function get(c: Context, body: GetDevice, apikey: Database['public'
       .eq('app_id', body.app_id)
       .eq('name', body.channel)
       .single()
-    if (dbError || !dataChannel)
+    if (dbError || !dataChannel) {
+      console.log('Cannot find version', dbError)
       return c.json({ status: 'Cannot find version', error: JSON.stringify(dbError) }, 400)
+    }
 
     return c.json(dataChannel)
   }
@@ -85,32 +89,58 @@ export async function get(c: Context, body: GetDevice, apikey: Database['public'
       .eq('app_id', body.app_id)
       .range(from, to)
       .order('created_at', { ascending: true })
-    if (dbError || !dataChannels || !dataChannels.length)
-      return c.json([])
+    if (dbError || !dataChannels) {
+      console.log('Cannot find channels', dbError)
+      return c.json({ status: 'Cannot find channels', error: JSON.stringify(dbError) }, 400)
+    }
     return c.json(dataChannels)
   }
 }
 
 export async function deleteChannel(c: Context, body: ChannelSet, apikey: Database['public']['Tables']['apikeys']['Row']): Promise<Response> {
-  if (!(await hasAppRight(c, body.app_id, apikey.user_id, 'admin')))
+  if (!(await hasAppRight(c, body.app_id, apikey.user_id, 'admin'))) {
+    console.log('You can\'t access this app', body.app_id)
     return c.json({ status: 'You can\'t access this app', app_id: body.app_id }, 400)
+  }
+  if (!body.channel) {
+    console.log('You must provide a channel name')
+    return c.json({ status: 'You must provide a channel name' }, 400)
+  }
 
   try {
-    const { error: dbError } = await supabaseAdmin(c)
+    // search if that exist first
+    const { data: dataChannel, error: dbError } = await supabaseAdmin(c)
+      .from('channels')
+      .select('id')
+      .eq('app_id', body.app_id)
+      .eq('name', body.channel)
+      .single()
+    if (dbError || !dataChannel) {
+      console.log('Cannot find channel', dbError)
+      return c.json({ status: 'Cannot find channel', error: JSON.stringify(dbError) }, 400)
+    }
+    await supabaseAdmin(c)
       .from('channels')
       .delete()
       .eq('app_id', body.app_id)
       .eq('name', body.channel)
-    if (dbError)
-      return c.json({ status: 'Cannot delete channels', error: JSON.stringify(dbError) }, 400)
   }
   catch (e) {
+    console.log('Cannot delete channels', e)
     return c.json({ status: 'Cannot delete channels', error: JSON.stringify(e) }, 500)
   }
   return c.json(BRES)
 }
 
 export async function post(c: Context, body: ChannelSet, apikey: Database['public']['Tables']['apikeys']['Row']): Promise<Response> {
+  const { data: org, error } = await supabaseAdmin(c).from('apps')
+    .select('owner_org')
+    .eq('app_id', body.app_id)
+    .single()
+  if (error || !org) {
+    console.log('Cannot find app', error)
+    return c.json({ status: 'Cannot find app', error: JSON.stringify(error) }, 400)
+  }
   const channel: Database['public']['Tables']['channels']['Insert'] = {
     created_by: apikey.user_id,
     app_id: body.app_id,
@@ -126,26 +156,49 @@ export async function post(c: Context, body: ChannelSet, apikey: Database['publi
     version: -1,
     owner_org: EMPTY_UUID,
   }
+
   if (body.version) {
     const { data, error: vError } = await supabaseAdmin(c)
       .from('app_versions')
       .select()
       .eq('app_id', body.app_id)
       .eq('name', body.version)
-      .eq('user_id', apikey.user_id)
+      .eq('owner_org', org.owner_org)
       .eq('deleted', false)
       .single()
-    if (vError || !data)
+    if (vError || !data) {
+      console.log('Cannot find unknown version', vError)
       return c.json({ status: 'Cannot find version', error: JSON.stringify(vError) }, 400)
+    }
 
     channel.version = data.id
   }
+  else {
+    // find the unknown version
+    const { data: dataVersion, error: dbError } = await supabaseAdmin(c)
+      .from('app_versions')
+      .select('id')
+      .eq('app_id', body.app_id)
+      .eq('owner_org', org.owner_org)
+      .eq('name', 'unknown')
+      .eq('deleted', true)
+      .single()
+    if (dbError || !dataVersion) {
+      console.log('Cannot find unknown version', dbError)
+      return c.json({ status: 'Cannot find version', error: JSON.stringify(dbError) }, 400)
+    }
+
+    channel.version = dataVersion.id
+  }
   try {
     const { error: dbError } = await updateOrCreateChannel(c, channel)
-    if (dbError)
+    if (dbError) {
+      console.log('Cannot create channel', dbError)
       return c.json({ status: 'Cannot create channel', error: JSON.stringify(dbError) }, 400)
+    }
   }
   catch (e) {
+    console.log('Cannot create channel', e)
     return c.json({ status: 'Cannot create channel', error: JSON.stringify(e) }, 500)
   }
   return c.json(BRES)
